requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.6.0
bleach==3.1.5
blinker==1.4
Brlapi==0.6.6
certifi==2019.9.11
chardet==3.0.4
click==6.7
colorama==0.3.7
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.1.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.2.8
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
entrypoints==0.3
Flask==0.12.2
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
importlib-metadata==1.6.1
imutils==0.5.3
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==0.24
itypes==1.1.0
jedi==0.17.0
Jinja2==2.10
joblib==0.15.1
jsonschema==3.2.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.4
netifaces==0.10.4
nose==1.3.7
numexpr==2.6.4
numpy==1.18.5
oauth==1.0.1
olefile==0.45.1
openapi-codec==1.3.2
opencv-python==4.2.0.32
packaging==20.4
pandas==1.0.4
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.2.1
pickleshare==0.7.5
Pillow==5.1.0
pipenv==2018.11.26
pluggy==0.6.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.0.0
psutil==5.4.2
psycopg2==2.7.4
ptyprocess==0.6.0
py==1.5.2
PyAudio==0.2.11
pycairo==1.16.2
PyChromecast==0.8.1
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pyrsistent==0.16.0
pytest==3.3.2
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==19.0.1
QtPy==1.9.0
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-learn==0.19.1
scipy==1.4.1
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.13.2
six==1.14.0
sklearn==0.0
sklearn-pandas==1.5.0
sqlparse==0.3.0
system-service==0.3
systemd-python==234
tables==3.4.2
terminado==0.8.3
testpath==0.4.4
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
typed-ast==1.4.1
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.22
usb-creator==0.3.3
virtualenv==16.7.8
virtualenv-clone==0.5.3
wadllib==1.3.2
wcwidth==0.2.3
webencodings==0.5
Werkzeug==0.14.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xkit==0.0.0
youtube-dl==2020.1.24
zeroconf==0.19.1
zipp==3.1.0
zope.interface==4.3.2
